Extract paint loop to screen_line.rs (refs #88)
Moves the large CallbackTrait::prepare (~800 lines of per-cell shaping
and instance emission), the paint() draw-submission method, and the
emit_custom_glyph helper out of callback.rs into a new screen_line.rs
module. Mirrors wezterm-gui's termwindow/render/screen_line.rs.

callback.rs is now a thin 18-line module holding only the
TerminalPaintCallback struct definition.

No behavior change — pure mechanical refactor.
